countByBaggage

+++ src/test/java/Test3.java
@@ -15,29 +15,29 @@ public class Test3 {
         String[] times = new String[]{"周一","周二","周三","周四","周五"};
         List<BaggageLineData> data = new ArrayList<>();
         data.add(
-                BaggageLineData.builder().time("周一").type("CA").count(10).build()
+                BaggageLineData.builder().time("周一").type("CA").count(10.0).build()
         );
         data.add(
-                BaggageLineData.builder().time("周五").type("CA").count(20).build()
+                BaggageLineData.builder().time("周五").type("CA").count(20.0).build()
         );
         data.add(
-                BaggageLineData.builder().time("周三").type("CU").count(20).build()
+                BaggageLineData.builder().time("周三").type("CU").count(20.0).build()
         );
 
 
-        Map<String, int[]> result = validate(times, data);
+        Map<String, Double[]> result = validate(times, data);
         System.out.println(result);
 
 
 
     }
 
-    private static Map<String, int[]> validate(String[] times, List<BaggageLineData> data) {
-        Map<String,int[]> result = new HashMap<>();
+    private static Map<String, Double[]> validate(String[] times, List<BaggageLineData> data) {
+        Map<String,Double[]> result = new HashMap<>();
         for(int i=0;i<data.size();i++){
             BaggageLineData baggage = data.get(i);
             if(result.get(baggage.getType())==null){
-                result.put(baggage.getType(),new int[times.length]);
+                result.put(baggage.getType(),new Double[times.length]);
             }
             for(int j=0;j<times.length;j++){
                 if(times[j].equals(baggage.getTime())){
